+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### 2022
 
-- v1.0.1 (2022-09-21)
+- v1.0.1 - v1.1.0 (2022-09-21)
   - optimize validate logic
+  - add methods: `OnlyContainPrefixWords`, `OnlyContainSuffixWords`, `MustNotContainPrefixWords`, `MustNotContainSuffixWords`
+  - add constants: `SPECIALCHARS`, `QUOTES`, `BRACKETS`, `OPERATORS`
 
 - v1.0.0 (2022-09-20)
     - Initial release